fix: update the getNetworks function in the bindings
