+++ sdk/src/core/wallet.rs
@@ -1005,7 +1005,7 @@ mod tests {
                     wallet_manager: Box::new(mock_wallet_manager),
                 });
                 sdk.access_token = Some(TOKEN.clone());
-                sdk.set_networks(Some(example_networks()));
+                sdk.set_networks(example_networks());
                 sdk.set_network(example_network_id(Currency::Iota)).await.unwrap();
 
                 let mock_request = SetUserAddressRequest {
@@ -1080,7 +1080,7 @@ mod tests {
                     username: USERNAME.into(),
                     wallet_manager: Box::new(mock_wallet_manager),
                 });
-                sdk.set_networks(Some(example_networks()));
+                sdk.set_networks(example_networks());
                 sdk.set_network(example_network_id(Currency::Iota)).await.unwrap();
             }
             Err(error) => {
@@ -1131,7 +1131,7 @@ mod tests {
                     username: USERNAME.into(),
                     wallet_manager: Box::new(mock_wallet_manager),
                 });
-                sdk.set_networks(Some(example_networks()));
+                sdk.set_networks(example_networks());
                 sdk.set_network(example_network_id(Currency::Iota)).await.unwrap();
             }
             Err(error) => {
@@ -1183,7 +1183,7 @@ mod tests {
                     username: USERNAME.into(),
                     wallet_manager: Box::new(mock_wallet_manager),
                 });
-                sdk.set_networks(Some(example_networks()));
+                sdk.set_networks(example_networks());
                 sdk.set_network(example_network_id(Currency::Iota)).await.unwrap();
             }
             Err(error) => {
